[#49][#52] Add support for scanning S3
Adds following command line options to allow for scanning S3 buckets:

- s3_endpoint_domain
- s3_region_name
- s3_keypair
- s3_proxy_url

s3_keypair must point to a file on the local machine with an S3 keypair.
s3_proxy_url is only needed if using a proxy to connect to the S3 endpoint.

S3 buckets are scanned via "chunking" of the retrieved objects such that
each task is delegated a chunk of objects, rather than just one. The
number of objects to process can be configured with the files_per_task
option. This value was defaulted to 50 based on benchmarking with 16 Celery
workers on one machine. Further benchmarking needed to determine a good
default as well as recommendations for different configurations and ingest
job types.

scandir is now the only option for scanning filesystems. As such, list_dir
and scan_dir_list command line options have been removed.