Make tests use 1 file per task
Tests were failing due to a set number of expected error messages.
This was based on the assumption that each task handled 1 path. With
the introduction of "chunking" and a default value of 50, the number
of error messages was coming out to 1 since, in most cases, the number
of files to process was less than the chunk size.