[#76] Add max redis version and requirements.txt
An upper limit has been placed on the Redis version to prevent
moving to the 3.x major version. This can be changed as needed,
but will prevent future versions of Redis from breaking existing
dependencies and their sub-dependencies.

A requirements.txt file has been added to communicate the exact
dependency versions that are supported by the currently released
version.