[#31] Handle invalid zone name in target coll
sync_task now raises an exception when the destination/target iRODS
collection contains an invalid zone name (that is, when '/zoneName'
collection does not exist), the target collection is root, or if root
does not exist as a collection. This will prevent scans with faulty
target collection arguments.

Added a note to the README about the requirement of iRODS authentication
in the both the Celery worker nodes and the node running the sync job.