[#109] Update docker steps for Celery
Provides two Dockerfiles:

1. Automated ingest as a python application image
2. Celery worker image based on ingest image

icommands.env and irods_envirionment.json files are required in and
outside of the container to make it run properly. This serves as the
connection info/authentication for our iRODS client.

Also provides a set of wrapper bash scripts for easy use of the
containers:

1. Scripts for listing, starting, stopping, and watching ingest jobs.
2. "Point-and-click" script: Minimizes typing by standing up Celery
   workers, scanning the target to the destination logical path using
   mainly default settings, and tearing everything down at the end.

NOTE: All of these scripts - especially point_and_click.sh - make a
series of assumptions about where they are running and what is
available to them at runtime. Please do not use these in production.