[#110] Add several interfaces for refactor
In order to increase extensibility, decouple features, and generally
improve the internal interfaces, this change adds a number of new
interfaces.

- custom_event_handler - Accessing methods and settings set inside of a
user-defined event handler module has been encapsulated in a class.

- redis_key - Interacting with different types of redis keys for
various purposes throughout the framework and the CLI has been
captured in a hierarchy of classes with accessors and mutators.

- scanner - This class abstracts the concepts of walking a list of
items and what it means to "sync" an item in order to hide the
details of implementation from the Celery tasks. This allows for more
easily adding additional scanner types in the future.

- sync_actions - These are the base actions used by the CLI which have
been captured inside of a separate module to separate them from the
Celery tasks. Cleaner separation of concerns.

- sync_job - The concept of an overall sync job has also been captured
in a class in order to hold context and provide a single interface
for tracking and managing individual tasks associated with a given
job.

- task_queue - The task queue abstracts the sync job from the Celery
tasks so that they can add tasks to the queue for other Celery
workers to pick up without having to know how it works. NOTE: This
class is not currently used as the separation of different concepts
into modules didn't quite work out, but we'll keep it here anyway.

The rest of the changes have to do with updating the existing Celery
tasks, CLI, and tests to use all of the new things.