[#86][#117] Test suite cleanup + docker image
This change cleans up the automated ingest test suite and ensures
that all tests are passing insofar as the server allows (Github issue
numbers referenced where relevant for skipping).

iRODS and redis server information is currently hard-coded for
the tests. iRODS client information is also hard-coded.

When fixing the UnicodeEncodeError tests, it appears that it is able
to create a collection in iRODS with a contained data object with no
name, leading to a data object to which one cannot refer. It must be
removed directly from the database. There is an issue in the main repo
for this: https://github.com/irods/irods/issues/4494