[#147][#157] Allow running workers with env only
$IRODS_ENVIRONMENT_FILE is still supported but if the env contains the
supported environment variables, this makes deployment of worker nodes
much easier. Here are the supported/expected environment variables:
    IRODS_HOST
    IRODS_PORT
    IRODS_USER_NAME
    IRODS_ZONE_NAME
    IRODS_PASSWORD

If any of these is missing, an environment file in JSON format is
expected to be at the path specified by $IRODS_ENVIRONMENT_FILE. The
environment file should contain at least the following information:
    {
        "irods_host": <string>,
        "irods_port": <int>,
        "irods_user": <string>,
        "irods_zone_name": <string>
    }

An .irodsA file is expected to be in the home directory of the Linux
user running the Celery workers on a given node in order to authenticate
with iRODS.

SSL environments still require the use of an irods_environment file, but
it is no longer required when SSL is not being used.