[#201] Amend test for post_job
The post_job test only runs one worker and passes because there is only
one worker to execute the post_job. The post_job has an issue where
multiple workers execute the post_job. This test opens the file in the
example post_job event_handler in append mode to demonstrate that the
write only occurs once. Additionally, 4 workers are used instead of 1
because the problem only manifests with more than 1 worker.