[#219] Add tests for pre/post event handler methods
This commit adds a series of tests for the pre/post event handler methods
for our various functions, including:

 - data_obj_create
 - data_obj_modify
 - coll_create
 - coll_modify

For each function, all Operations are configured in a custom event handler
which trigger said functions and ensures that the pre and post methods are
fired appropriately.

Operation.REGISTER_AS_REPLICA_SYNC requires special tests, so this operation
is being skipped. Operation.PUT is skipped for the modify functions because it
does not do anything on updates. Operation.NO_OP obviously does not invoke any
of these functions, so it is being skipped for all the tests.