[#248] docker: Remove performance_testing project
Also added a section in the README of the demo project for how to use
it to test out various configurations for performance numbers. As
stated in the README, Docker isn't the best place to go for performance
numbers, but it can give an idea of the effect of various configurations
in various environments.