[#211] Create separate tasks for filesystem and S3
This commit creates an alternative implementation of the ingest Celery
app. The most relevant changes to the project structure are described
below.

The Celery application is now housed in celery.py which allows the Celery
application run by the workers to refer simply to the name of the Python
module irods_capability_automated_ingest instead of requiring
irods_capability_automated_ingest.sync_task. The Celery application includes
the tasks defined in the tasks subdirectory.

The tasks subdirectory includes a file for syncing a filesystem and a file
for syncing an S3 bucket. There is currently a lot of duplicate code between
the two, but separating these is an important step for adding more types of
storage for syncing.

In order to remove circular dependencies, an irods subdirectory has also
been added to provide the functionality for interactions with PRC. The Celery
tasks call into these functions just like they always had, but, as they always
have, storage-specific implementations are needed in order to implement things
like reading data and performing parallel transfers.