[#289] Fix README description for PUT operation
The README says that PUT will emit an error if the target data object
already exists. This is not true. PUT ignores the existing data object
and does not sync any data.

This commit also adds a more explicit if-else case in the relevant
places in the code to make it clearer when this comes into play.