[#124] Add test for deep, nonexistent subcollection
If a deep, nonexistent subcollection is the root target
collection for a sync job and the number of path elements
exceeds the number of path elements in the source path,
the sync job could get stuck waiting on a redis_lock.

This change adds a test for a filesystem sync and an S3
bucket sync which demonstrates the behavior.