[#132] Add Minio service to ingest-test, remove Minio volume
This commit adds a Minio service to ingest-test so that S3 tests can be
run using this Compose project.

This commit removes the Docker volume used with the Minio service in
the docker/demo Compose project. The volume is not needed because neither
the ingest workers nor iRODS need access to the underlying storage as
everything goes through Minio.