[#210] Track jobs stopped by irods_sync stop
Adds a new list to the tracked Redis keys:
irods_ingest_stopped_jobs. This tracks jobs which
are manually stopped with the stop subcommand.

Displays a warning message to the irods_sync script's
log output when a job being monitored by start --progress
or watch is stopped.

The list subcommand now shows the list of stopped jobs
and last-known information about them when they were stopped.