[#91,#92,#210] Document irods_sync script subcommands
This commit documents several new irods_sync script features
as well as adds documentation for some of the subcommands
which previously had little to no documentation.

The stop subcommand now adds job names to a new Redis key,
which is being documented.

The list subcommand got some new information to display in
addition to being able to list stopped jobs.

The irods_sync script section has been reorganized as well
so that the parts fit together a little better.