[#299] v0.6.0, plus python version bumps
The minimum required python version has been bumped
to 3.8 and the ingest version has been bumped to 0.6.0.

The following words capture what we are trying to
accomplish with these version listings:

python_requires reflects the minimum python version
required to run the application. This number should not
be lower than the lowest supported version of python
required to run the version of python-irodsclient upon
which this application depends.

Programming Languages lists every python version from
the minimum value of python_requires all the way up to
the latest release of python which can run the application.
This number should not be higher than the highest supported
version of python required to run the version of
python-irodsclient upon which this application depends.